restructured examples
