+++ examples/src/main/scala/com/example/Warehouse.scala
