+++ examples/src/test/scala/com/example/HigherOrderFunctionsTest.scala
@@ -1,15 +1,15 @@
 // Copyright (c) 2011-2015 ScalaMock Contributors (https://github.com/paulbutcher/ScalaMock/graphs/contributors)
-// 
+//
 // Permission is hereby granted, free of charge, to any person obtaining a copy
 // of this software and associated documentation files (the "Software"), to deal
 // in the Software without restriction, including without limitation the rights
 // to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
 // copies of the Software, and to permit persons to whom the Software is
 // furnished to do so, subject to the following conditions:
-// 
+//
 // The above copyright notice and this permission notice shall be included in
 // all copies or substantial portions of the Software.
-// 
+//
 // THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
 // IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
 // FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
@@ -18,31 +18,28 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
 
-package com.example.function.mockitostyle
+package com.example
 
 import org.scalamock.scalatest.MockFactory
 import org.scalatest.freespec.AnyFreeSpec
 
 class HigherOrderFunctionsTest extends AnyFreeSpec with MockFactory {
+
   import language.postfixOps
 
   "HigherOrderFunctionsTest" - {
 
     "testMap" in {
-      val f = stubFunction[Int, String]
+      val f = mockFunction[Int, String]
 
-      f when (1) returns "one"
-      f when (2) returns "two"
-      f when (3) returns "three"
+      inSequence {
+        f expects (1) returning "one" once;
+        f expects (2) returning "two" once;
+        f expects (3) returning "three" once;
+      }
 
       assertResult(Seq("one", "two", "three")) {
         Seq(1, 2, 3) map f
-      }
-
-      inSequence {
-        f verify (1) once;
-        f verify (2) once;
-        f verify (3) once;
       }
     }
 
@@ -52,32 +49,26 @@ class HigherOrderFunctionsTest extends AnyFreeSpec with MockFactory {
           what
       }
 
-      val f = stubFunction[Unit]
+      val f = mockFunction[Unit]
+      f.expects().repeated(4).times()
 
       repeat(4) {
         f()
       }
-
-      f.verify().repeated(4).times()
     }
 
     "testFoldLeft" in {
-      val f = stubFunction[String, Int, String]
+      val f = mockFunction[String, Int, String]
 
-      f when("initial", 0) returns "intermediate one"
-      f when("intermediate one", 1) returns "intermediate two"
-      f when("intermediate two", 2) returns "intermediate three"
-      f when("intermediate three", 3) returns "final"
+      inSequence {
+        f expects("initial", 0) returning "intermediate one" once;
+        f expects("intermediate one", 1) returning "intermediate two" once;
+        f expects("intermediate two", 2) returning "intermediate three" once;
+        f expects("intermediate three", 3) returning "final" once;
+      }
 
       assertResult("final") {
         Seq(0, 1, 2, 3).foldLeft("initial")(f)
-      }
-
-      inSequence {
-        f verify("initial", 0) once;
-        f verify("intermediate one", 1) once;
-        f verify("intermediate two", 2) once;
-        f verify("intermediate three", 3) once;
       }
     }
   }
