+++ examples/src/test/scala/com/example/mockitostyle/ControllerTest.scala
@@ -1,15 +1,15 @@
 // Copyright (c) 2011-2015 ScalaMock Contributors (https://github.com/paulbutcher/ScalaMock/graphs/contributors)
-// 
+//
 // Permission is hereby granted, free of charge, to any person obtaining a copy
 // of this software and associated documentation files (the "Software"), to deal
 // in the Software without restriction, including without limitation the rights
 // to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
 // copies of the Software, and to permit persons to whom the Software is
 // furnished to do so, subject to the following conditions:
-// 
+//
 // The above copyright notice and this permission notice shall be included in
 // all copies or substantial portions of the Software.
-// 
+//
 // THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
 // IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
 // FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
@@ -18,43 +18,37 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
 
-package com.example.makro
+package com.example.mockitostyle
 
+import com.example.{Controller, Turtle}
 import org.scalamock.scalatest.MockFactory
-import com.example.{Order, Warehouse}
-import org.scalatest.wordspec.AnyWordSpec
+import org.scalatest.funsuite.AnyFunSuite
 
-// This is a reworked version of the example from Martin Fowler's article
-// Mocks Aren't Stubs http://martinfowler.com/articles/mocksArentStubs.html
-class OrderTest extends AnyWordSpec with MockFactory {
-  import language.postfixOps
-  
-  "An order" when {
-    "in stock" should {
-      "remove inventory" in {
-        val mockWarehouse = mock[Warehouse]
-        inSequence {
-          (mockWarehouse.hasInventory _) expects ("Talisker", 50) returning true
-          (mockWarehouse.remove _) expects ("Talisker", 50) once
-        }
-        
-        val order = new Order("Talisker", 50)
-        order.fill(mockWarehouse)
-        
-        assert(order.isFilled)
+import scala.math.{Pi, sqrt}
+ 
+class ControllerTest extends AnyFunSuite with MockFactory {
+  import scala.language.postfixOps
+ 
+  test("draw line") {
+    val mockTurtle = stub[Turtle]
+    val controller = new Controller(mockTurtle)
+
+    inSequence {
+      inAnyOrder {
+        (mockTurtle.getPosition _).when().returns(0.0, 0.0)
+        (mockTurtle.getAngle _).when().returns(0.0).once()
       }
+      (mockTurtle.getAngle _).when().returns(Pi / 4)
     }
+ 
+    controller.drawLine((1.0, 1.0), (2.0, 1.0))
     
-    "out of stock" should {
-      "remove nothing" in {
-        val mockWarehouse = mock[Warehouse]
-        (mockWarehouse.hasInventory _) stubs (*, *) returning false
-        
-        val order = new Order("Talisker", 50)
-        order.fill(mockWarehouse)
-        
-        assert(!order.isFilled)
-      }
+    inSequence {
+      (mockTurtle.turn _).verify(~(Pi / 4))
+      (mockTurtle.forward _).verify(~sqrt(2.0))
+      (mockTurtle.turn _).verify(~(-Pi / 4))
+      (() => mockTurtle.penDown()).verify()
+      (mockTurtle.forward _).verify(1.0)
     }
   }
 }
