+++ examples/src/test/scala/com/example/mockitostyle/HigherOrderFunctionsTest.scala
@@ -18,28 +18,31 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
 
-package com.example.function
+package com.example.mockitostyle
 
 import org.scalamock.scalatest.MockFactory
 import org.scalatest.freespec.AnyFreeSpec
 
 class HigherOrderFunctionsTest extends AnyFreeSpec with MockFactory {
-
   import language.postfixOps
 
   "HigherOrderFunctionsTest" - {
 
     "testMap" in {
-      val f = mockFunction[Int, String]
+      val f = stubFunction[Int, String]
 
-      inSequence {
-        f expects (1) returning "one" once;
-        f expects (2) returning "two" once;
-        f expects (3) returning "three" once;
-      }
+      f when (1) returns "one"
+      f when (2) returns "two"
+      f when (3) returns "three"
 
       assertResult(Seq("one", "two", "three")) {
         Seq(1, 2, 3) map f
+      }
+
+      inSequence {
+        f verify (1) once;
+        f verify (2) once;
+        f verify (3) once;
       }
     }
 
@@ -49,26 +52,32 @@ class HigherOrderFunctionsTest extends AnyFreeSpec with MockFactory {
           what
       }
 
-      val f = mockFunction[Unit]
-      f.expects().repeated(4).times()
+      val f = stubFunction[Unit]
 
       repeat(4) {
         f()
       }
+
+      f.verify().repeated(4).times()
     }
 
     "testFoldLeft" in {
-      val f = mockFunction[String, Int, String]
+      val f = stubFunction[String, Int, String]
 
-      inSequence {
-        f expects("initial", 0) returning "intermediate one" once;
-        f expects("intermediate one", 1) returning "intermediate two" once;
-        f expects("intermediate two", 2) returning "intermediate three" once;
-        f expects("intermediate three", 3) returning "final" once;
-      }
+      f when("initial", 0) returns "intermediate one"
+      f when("intermediate one", 1) returns "intermediate two"
+      f when("intermediate two", 2) returns "intermediate three"
+      f when("intermediate three", 3) returns "final"
 
       assertResult("final") {
         Seq(0, 1, 2, 3).foldLeft("initial")(f)
+      }
+
+      inSequence {
+        f verify("initial", 0) once;
+        f verify("intermediate one", 1) once;
+        f verify("intermediate two", 2) once;
+        f verify("intermediate three", 3) once;
       }
     }
   }
